refactor(ui): remove clutter
First refactor pass, removing clutter
