--- crates/ui/src/layout/layers.rs
@@ -1,8 +1,0 @@
-//! Contains the rendering logic for the [`crate::layout::EditorPanels::Layers`] pane.
-use crate::layout::EditorLayout;
-use egui::Ui;
-
-/// Handles the rendering of the [`crate::layout::EditorPanels::Layers`] tab.
-pub(super) fn render(_viewer: &mut EditorLayout, ui: &mut Ui) {
-    ui.label("Rendering layers here");
-}
