+++ crates/ui/src/lib.rs
@@ -1,7 +1,6 @@
 #![doc = include_str!("../README.md")]
 
 mod camera;
-mod dialogs;
 mod layout;
 mod notifications;
 mod plugin;
